// orbit-frontend/src/lib/api.ts
export type WantItem = {
  id: number
  album_id: number | null
  album_title: string | null
  album_cover: string | null
  artist_id: number | null
  artist_name: string | null
  track_id: number | null
  track_title: string | null
  created_at: string
}